requirements.txt
oracledb
requests
bcrypt
python-dotenv